refactor: simulation class for logic and data
The mandatory part of the star refactor is finished ! The
Simulation class becomes the center of the drake simulation. This
is were stars are created and new civlizations are born. All that
is really used in the World classes is the new civilizations spawn
to create the signal bubbles. Everything becomes strictly
graphical there.