fix: save button works on first try
Just removed useless returnValue test. It does not serve any purpose
now since the "SaveStatus" is the return value of the saveDrakeForm
function and that the cancel button is not longer a submit with a
"cancel" value but a reset button. Submit is necessarily a "save"
click action from the user so there is no need to check the dialog's
returnValue (which was not the right method anyway).