fix: add a reduction maximum
I did not think about that, but it is actually a problem to reduce
the bubbles size too much even when limiting their number to a
reasonably low value. The simulation slows down significantly
anyway because of the accelarated rate of bubble creation and
destruction.

The key here seems to adjust the bubble rate. If a bubble rate of
lets say 5 per second is implemented the bubbles wont all appear
and pop at the same time, it will be smoothed out. Also this should
have the added benefit of adding diversity to the output (bubble
will have different sizes) which seems more natural. The bubble
rate also should not change anything for the lowest settings where
the civilization yearly spawn rate is way lower.