fix: huge error in the spawn function
The rate was not computed correctly. Only multiply the spawnRate by
the elasped time if this value is less than one (less than a year).
Because the computed value is re-used for each year elapsed so the
rate has to be yearly.

For example if elapsed was equal to a 100 and spawnRate to 0.1,
then rate would have been equal to 100 * 0.1 = 10. So it would have
been 10 * 100 = 1000 civilizations born in a hundred year instead of
10. So here the spawnRate can be used directly.