refactor: optimize simulation _kill function
Do not set lifetime at civilization creation instead compute a
killRate based on the current spawnRate and number of civilizations.
This way it is not useful to keep living civilizations in any
specific order. More efficient.

A byproduct of this change is that the effective lifetime is almost
constant because the _kill function does not randomly selects the
civilizations that should die. Instead it removes the death count
elements from the beginning living civilzations array, effectively
killing the older civilizations first (since the _spawn function simply
pushes to the living array when adding a civilization). This is a bit
boring and unrealistic but it is efficient.