fix: only red bubbles on fast reasonable preset
This was because the first and last signal were created at the same time
(one tick was long enough for the civilization to be born and die). The
problem was that at each tick every event used the same 'date' as a
starting point (only delta * speed = elapsed was given to the tick
functions). Now the actual birth and death times are used to create the
bubbles.