refactor: split config in 3
The first one is the configTemplate (limits and starting default
values), the second one is configDefaults which stores the current
config default values (since they now can be changed - at least for the
preset setting - they are also in localStorage like the presets) and the
third is config, the actual current configuration. The values are
directly stored as key/value pairs, so no need for the annoying
config[_name_].current syntax, just use config[_name_] now.