fix: click handler for touchscreens
The mouse position is reset on click. Otherwise, since mousemove is
obviously never fired when there is no mouse, the mouse position was
never changed. So a click somewhere on the screen just remained there
and interacted with the elements as if it was always happening (because
of the Signals' onMouseMove method that is called at each tick).